+++ src/EventStore.Core.Tests/ClientAPI/event_store_connection_should.cs
@@ -8,9 +8,9 @@ using NUnit.Framework;
 namespace EventStore.Core.Tests.ClientAPI {
 	[Category("ClientAPI"), Category("LongRunning")]
 	[TestFixture(typeof(LogFormat.V2), typeof(string), TcpType.Normal)]
-	[TestFixture(typeof(LogFormat.V3), typeof(long), TcpType.Normal)]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint), TcpType.Normal)]
 	[TestFixture(typeof(LogFormat.V2), typeof(string), TcpType.Ssl)]
-	[TestFixture(typeof(LogFormat.V3), typeof(long), TcpType.Ssl)]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint), TcpType.Ssl)]
 	public class event_store_connection_should<TLogFormat, TStreamId> : SpecificationWithDirectoryPerTestFixture {
 		private readonly TcpType _tcpType;
 		private MiniNode<TLogFormat, TStreamId> _node;
