+++ src/EventStore.Core.Tests/Services/Storage/AllReader/when_multiple_single_writes_are_after_transaction_end_but_before_commit_is_present.cs
@@ -4,7 +4,7 @@ using NUnit.Framework;
 
 namespace EventStore.Core.Tests.Services.Storage.AllReader {
 	[TestFixture(typeof(LogFormat.V2), typeof(string))]
-	[TestFixture(typeof(LogFormat.V3), typeof(long), Ignore = "Explicit transactions are not supported yet by Log V3")]
+	[TestFixture(typeof(LogFormat.V3), typeof(uint), Ignore = "Explicit transactions are not supported yet by Log V3")]
 	public class
 		when_multiple_single_writes_are_after_transaction_end_but_before_commit_is_present<TLogFormat, TStreamId>
 		: RepeatableDbTestScenario<TLogFormat, TStreamId> {
